discord/requirements.txt
chromadb==0.5.0
config==0.5.1
googletrans==3.0.0
langchain==0.2.7
langchain_anthropic==0.1.19
langchain_community==0.2.7
langchain_core==0.2.13
langgraph==0.1.7
matplotlib==3.9.1
numpy==2.0.0
pandas==2.2.2
playwright==1.45.0
pydantic==2.8.2
python-dotenv==1.0.1
Requests==2.32.3
